+++ src/components/coding-activity.tsx
@@ -8,8 +8,7 @@ import { useWakatime } from '@/hooks'
 import cn from '@/lib/cn'
 import { formatDate } from '@/lib/utils'
 
-import { Spinner } from './common'
-import OverviewCard from './common/overview-card'
+import { OverviewCard, Spinner } from './common'
 import Progress from './common/progress'
 import { WakaTime as WakaTimeIcon } from './icons'
 import Link from './link'
