+++ src/components/common/index.ts
@@ -5,6 +5,7 @@ import Divider from './divider'
 import EmptyState from './empty-state'
 import IncrementCounter from './increment-counter'
 import Input from './input'
+import OverviewCard from './overview-card'
 import Section from './section'
 import Spinner from './spinner'
 import Tooltip from './tooltip'
@@ -17,6 +18,7 @@ export {
   EmptyState,
   IncrementCounter,
   Input,
+  OverviewCard,
   Section,
   Spinner,
   Tooltip,
