+++ src/components/contribution-overview.tsx
@@ -1,7 +1,7 @@
 import cn from '@/lib/cn'
 import type { ContributionCalendar } from '@/types/github'
 
-import OverviewCard from './common/overview-card'
+import { OverviewCard } from './common'
 
 interface ContributionOverviewProps {
   data?: ContributionCalendar
